Refactor: Rename FrameDiffer => DetectMotion
It's really the core of the detect_motion algorithm, but in an API
that's more convenient to use from press_and_wait.
